moved events service to server folder
