--- events/.gitignore
@@ -1,2 +1,0 @@
-venv/
-__pycache__/
