+++ server/events.py
@@ -1,0 +1,5 @@
+from events import main
+import asyncio
+
+if __name__ == "__main__":
+    asyncio.run(main())
