+++ server/events/main.py
@@ -1,8 +1,7 @@
-import asyncio
 import json
-from websockets.asyncio.server import serve # type: ignore
+from websockets.asyncio.server import serve
 
-from meeting_handler import MeetingHandler
+from .meeting_handler import MeetingHandler
 
 async def handler(websocket):
     handlers = {
@@ -30,6 +29,3 @@ async def main():
     async with serve(handler, "", 8001) as server:
         print("Server running on ws://localhost:8001")
         await server.serve_forever()
-
-if __name__ == "__main__":
-    asyncio.run(main())
