+++ server/requirements.txt
@@ -5,3 +5,4 @@ psycopg[binary]
 pytest
 reportlab
 firebase-admin
+websockets
